Move and improve Manifist managers
