+++ src/lib/manifest/index.ts
@@ -1,0 +1,2 @@
+export { ManifestEditor } from "./ManifestEditor";
+export { ManifestValidator } from "./ManifestValidator";
