+++ src/modules/apps/link/Link.ts
@@ -1,0 +1,7 @@
+export class Link {
+  constructor() {}
+
+  public start() {}
+
+  private performInitialLink() {}
+}
